--- mariaDBConn.js
@@ -1,30 +1,0 @@
-import mariadb from "mariadb";
-import "dotenv/config";
-
-
-
-
-const pool = mariadb.createPool({
-    host: process.env.DBHOST,
-    port: process.env.DBPORT,
-    user: process.env.DBUSER,
-    password: process.env.DBPASS,
-    connectionLimit: 5
-})
-
-const GetUserList = async() => {
-    let conn, rows;
-    try{
-        conn = await pool.getConnection();
-        conn.query('USE nodejs_test');
-        rows = await conn.query('SELECT * FROM users');
-    }catch(err){
-        throw err;
-    }
-    finally{
-        if (conn) conn.end();
-        return rows[0];
-    }
-}
-
-export { GetUserList as getUserList };
